refactor(core): split 4 oversized files to comply with 500 NLOC limit
- cost_model.rs (737→297): extract cost_factors.rs (278) and calibration.rs (263)
- filter.rs (712→354): extract filter_array.rs (192) and filter_geo.rs (183)
- explain.rs (708→module): split into explain/mod.rs, types.rs (302), plan_builder.rs (401)
- responses.rs (615→443): extract responses_explain.rs (181)

All files now under 500 NLOC. Zero public API changes — re-exports preserve
backward compatibility. All existing tests pass.
