+++ crates/velesdb-core/src/api_types/mod.rs
@@ -6,6 +6,7 @@
 
 mod requests;
 mod responses;
+mod responses_explain;
 pub mod serde_id;
 #[cfg(test)]
 mod tests;
