+++ crates/velesdb-core/src/column_store/filter.rs
@@ -1,64 +1,15 @@
-//! Filter operations for ColumnStore.
+//! Scalar filter operations for `ColumnStore`.
 //!
-//! This module provides efficient filtering methods for column-oriented data,
-//! including bitmap-based operations for large datasets.
+//! This module provides efficient filtering methods for scalar column types
+//! (integer, string) including bitmap-based operations for large datasets.
+//!
+//! Array filters live in [`super::filter_array`], geo filters in
+//! [`super::filter_geo`].
 
 use roaring::RoaringBitmap;
 
-use super::haversine::haversine_distance;
 use super::types::{StringId, TypedColumn};
 use super::ColumnStore;
-
-/// Comparison operator for geo-distance filters.
-///
-/// Mirrors `velesql::CompareOp` but lives in the column-store layer
-/// to avoid a dependency on the parser module.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum CompareOp {
-    /// Equal (=)
-    Eq,
-    /// Not equal (!=)
-    NotEq,
-    /// Greater than (>)
-    Gt,
-    /// Greater than or equal (>=)
-    Gte,
-    /// Less than (<)
-    Lt,
-    /// Less than or equal (<=)
-    Lte,
-}
-
-/// Parameters for `GEO_DISTANCE` filter (>3 params → use struct per project rules).
-#[derive(Debug)]
-pub struct GeoDistanceParams<'a> {
-    /// Column name containing `GeoPoint` data.
-    pub column: &'a str,
-    /// Reference latitude in degrees.
-    pub lat: f64,
-    /// Reference longitude in degrees.
-    pub lng: f64,
-    /// Comparison operator to apply against the computed distance.
-    pub operator: CompareOp,
-    /// Distance threshold in meters.
-    pub threshold: f64,
-}
-
-/// Parameters for `GEO_BBOX` filter (>3 params → use struct per project rules).
-#[derive(Debug)]
-pub struct GeoBboxParams<'a> {
-    /// Column name containing `GeoPoint` data.
-    pub column: &'a str,
-    /// Minimum latitude of the bounding box.
-    pub lat_min: f64,
-    /// Minimum longitude of the bounding box.
-    pub lng_min: f64,
-    /// Maximum latitude of the bounding box.
-    pub lat_max: f64,
-    /// Maximum longitude of the bounding box.
-    pub lng_max: f64,
-}
 
 // =========================================================================
 // Private scan helpers — eliminate duplication across Vec / bitmap / count
@@ -399,314 +350,5 @@ impl ColumnStore {
                 _ => None,
             })
             .collect()
-    }
-
-    // =========================================================================
-    // Array CONTAINS Filtering
-    // =========================================================================
-
-    /// Returns row indices where the array column contains `value`.
-    ///
-    /// Returns empty results for non-existent or non-array columns.
-    /// Excludes deleted rows and null arrays.
-    #[must_use]
-    pub fn filter_contains(&self, column: &str, value: &super::types::ColumnValue) -> Vec<usize> {
-        let Some(TypedColumn::Array { data, .. }) = self.columns.get(column) else {
-            return Vec::new();
-        };
-        data.iter()
-            .enumerate()
-            .filter_map(|(idx, row)| {
-                let arr = row.as_ref()?;
-                if arr.contains(value) && !self.deleted_rows.contains(&idx) {
-                    Some(idx)
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Returns row indices where the array contains at least one of `values`.
-    ///
-    /// Uses `FxHashSet` optimization when M > 8 values.
-    /// Returns empty results for non-existent or non-array columns.
-    #[must_use]
-    pub fn filter_contains_any(
-        &self,
-        column: &str,
-        values: &[super::types::ColumnValue],
-    ) -> Vec<usize> {
-        let Some(TypedColumn::Array { data, .. }) = self.columns.get(column) else {
-            return Vec::new();
-        };
-        if values.len() > 8 {
-            self.contains_any_hashset(data, values)
-        } else {
-            self.contains_any_linear(data, values)
-        }
-    }
-
-    /// Linear scan for `filter_contains_any` with small value lists.
-    fn contains_any_linear(
-        &self,
-        data: &[Option<smallvec::SmallVec<[super::types::ColumnValue; 8]>>],
-        values: &[super::types::ColumnValue],
-    ) -> Vec<usize> {
-        data.iter()
-            .enumerate()
-            .filter_map(|(idx, row)| {
-                let arr = row.as_ref()?;
-                if values.iter().any(|v| arr.contains(v)) && !self.deleted_rows.contains(&idx) {
-                    Some(idx)
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// `FxHashSet`-based scan for `filter_contains_any` with large value lists.
-    fn contains_any_hashset(
-        &self,
-        data: &[Option<smallvec::SmallVec<[super::types::ColumnValue; 8]>>],
-        values: &[super::types::ColumnValue],
-    ) -> Vec<usize> {
-        let value_set: Vec<&super::types::ColumnValue> = values.iter().collect();
-        data.iter()
-            .enumerate()
-            .filter_map(|(idx, row)| {
-                let arr = row.as_ref()?;
-                if arr.iter().any(|e| value_set.contains(&e)) && !self.deleted_rows.contains(&idx) {
-                    Some(idx)
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Returns row indices where the array contains every value in `values`.
-    ///
-    /// Returns empty results for non-existent or non-array columns.
-    #[must_use]
-    pub fn filter_contains_all(
-        &self,
-        column: &str,
-        values: &[super::types::ColumnValue],
-    ) -> Vec<usize> {
-        let Some(TypedColumn::Array { data, .. }) = self.columns.get(column) else {
-            return Vec::new();
-        };
-        data.iter()
-            .enumerate()
-            .filter_map(|(idx, row)| {
-                let arr = row.as_ref()?;
-                if values.iter().all(|v| arr.contains(v)) && !self.deleted_rows.contains(&idx) {
-                    Some(idx)
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Bitmap variant of `filter_contains`.
-    ///
-    /// Safely skips indices exceeding `u32::MAX`.
-    #[must_use]
-    pub fn filter_contains_bitmap(
-        &self,
-        column: &str,
-        value: &super::types::ColumnValue,
-    ) -> RoaringBitmap {
-        let Some(TypedColumn::Array { data, .. }) = self.columns.get(column) else {
-            return RoaringBitmap::new();
-        };
-        data.iter()
-            .enumerate()
-            .filter_map(|(idx, row)| {
-                let arr = row.as_ref()?;
-                if arr.contains(value) && !self.deleted_rows.contains(&idx) {
-                    u32::try_from(idx).ok()
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Bitmap variant of `filter_contains_any`.
-    ///
-    /// Safely skips indices exceeding `u32::MAX`.
-    #[must_use]
-    pub fn filter_contains_any_bitmap(
-        &self,
-        column: &str,
-        values: &[super::types::ColumnValue],
-    ) -> RoaringBitmap {
-        let Some(TypedColumn::Array { data, .. }) = self.columns.get(column) else {
-            return RoaringBitmap::new();
-        };
-        data.iter()
-            .enumerate()
-            .filter_map(|(idx, row)| {
-                let arr = row.as_ref()?;
-                if values.iter().any(|v| arr.contains(v)) && !self.deleted_rows.contains(&idx) {
-                    u32::try_from(idx).ok()
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Bitmap variant of `filter_contains_all`.
-    ///
-    /// Safely skips indices exceeding `u32::MAX`.
-    #[must_use]
-    pub fn filter_contains_all_bitmap(
-        &self,
-        column: &str,
-        values: &[super::types::ColumnValue],
-    ) -> RoaringBitmap {
-        let Some(TypedColumn::Array { data, .. }) = self.columns.get(column) else {
-            return RoaringBitmap::new();
-        };
-        data.iter()
-            .enumerate()
-            .filter_map(|(idx, row)| {
-                let arr = row.as_ref()?;
-                if values.iter().all(|v| arr.contains(v)) && !self.deleted_rows.contains(&idx) {
-                    u32::try_from(idx).ok()
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    // =========================================================================
-    // GeoPoint Filtering
-    // =========================================================================
-
-    /// Returns row indices where the Haversine distance satisfies the comparison.
-    ///
-    /// Returns empty results for non-existent or non-GeoPoint columns.
-    /// Excludes deleted rows and null values.
-    #[must_use]
-    pub fn filter_geo_distance(&self, params: &GeoDistanceParams<'_>) -> Vec<usize> {
-        let Some(TypedColumn::GeoPoint(col)) = self.columns.get(params.column) else {
-            return Vec::new();
-        };
-        col.iter()
-            .enumerate()
-            .filter_map(|(idx, v)| {
-                let (lat, lng) = (*v).as_ref()?;
-                let dist = haversine_distance(*lat, *lng, params.lat, params.lng);
-                if compare_f64(dist, params.threshold, params.operator)
-                    && !self.deleted_rows.contains(&idx)
-                {
-                    Some(idx)
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Bitmap variant of `filter_geo_distance`.
-    ///
-    /// Safely skips indices exceeding `u32::MAX`.
-    #[must_use]
-    pub fn filter_geo_distance_bitmap(&self, params: &GeoDistanceParams<'_>) -> RoaringBitmap {
-        let Some(TypedColumn::GeoPoint(col)) = self.columns.get(params.column) else {
-            return RoaringBitmap::new();
-        };
-        col.iter()
-            .enumerate()
-            .filter_map(|(idx, v)| {
-                let (lat, lng) = (*v).as_ref()?;
-                let dist = haversine_distance(*lat, *lng, params.lat, params.lng);
-                if compare_f64(dist, params.threshold, params.operator)
-                    && !self.deleted_rows.contains(&idx)
-                {
-                    u32::try_from(idx).ok()
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Returns row indices where the GeoPoint falls within the bounding box (inclusive).
-    ///
-    /// Returns empty results for non-existent or non-GeoPoint columns,
-    /// or when `lat_min > lat_max` or `lng_min > lng_max`.
-    #[must_use]
-    pub fn filter_geo_bbox(&self, params: &GeoBboxParams<'_>) -> Vec<usize> {
-        if params.lat_min > params.lat_max || params.lng_min > params.lng_max {
-            return Vec::new();
-        }
-        let Some(TypedColumn::GeoPoint(col)) = self.columns.get(params.column) else {
-            return Vec::new();
-        };
-        col.iter()
-            .enumerate()
-            .filter_map(|(idx, v)| {
-                let (lat, lng) = (*v).as_ref()?;
-                if *lat >= params.lat_min
-                    && *lat <= params.lat_max
-                    && *lng >= params.lng_min
-                    && *lng <= params.lng_max
-                    && !self.deleted_rows.contains(&idx)
-                {
-                    Some(idx)
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Bitmap variant of `filter_geo_bbox`.
-    ///
-    /// Safely skips indices exceeding `u32::MAX`.
-    #[must_use]
-    pub fn filter_geo_bbox_bitmap(&self, params: &GeoBboxParams<'_>) -> RoaringBitmap {
-        if params.lat_min > params.lat_max || params.lng_min > params.lng_max {
-            return RoaringBitmap::new();
-        }
-        let Some(TypedColumn::GeoPoint(col)) = self.columns.get(params.column) else {
-            return RoaringBitmap::new();
-        };
-        col.iter()
-            .enumerate()
-            .filter_map(|(idx, v)| {
-                let (lat, lng) = (*v).as_ref()?;
-                if *lat >= params.lat_min
-                    && *lat <= params.lat_max
-                    && *lng >= params.lng_min
-                    && *lng <= params.lng_max
-                    && !self.deleted_rows.contains(&idx)
-                {
-                    u32::try_from(idx).ok()
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-}
-
-/// Applies a comparison operator to two `f64` values.
-fn compare_f64(a: f64, b: f64, op: CompareOp) -> bool {
-    match op {
-        CompareOp::Eq => (a - b).abs() < f64::EPSILON,
-        CompareOp::NotEq => (a - b).abs() >= f64::EPSILON,
-        CompareOp::Gt => a > b,
-        CompareOp::Gte => a >= b,
-        CompareOp::Lt => a < b,
-        CompareOp::Lte => a <= b,
     }
 }
