--- crates/velesdb-core/src/velesql/explain.rs
@@ -1,708 +1,0 @@
-//! Query plan explanation for `VelesQL`.
-//!
-//! This module provides EXPLAIN functionality to display query execution plans.
-//!
-//! # Example
-//!
-//! ```ignore
-//! use velesdb_core::velesql::{Parser, QueryPlan};
-//!
-//! let query = Parser::parse("SELECT * FROM docs WHERE vector NEAR $v LIMIT 10")?;
-//! let plan = QueryPlan::from_select(&query.select);
-//! println!("{}", plan.to_tree());
-//! ```
-
-mod formatter;
-mod node_stats;
-
-pub use node_stats::build_leaf_node_stats;
-
-use serde::{Deserialize, Serialize};
-use std::collections::HashSet;
-
-use super::ast::{Condition, LetBinding, SelectStatement};
-use crate::collection::query_cost::cost_model::OperationCostFactors;
-use crate::collection::search::query::match_planner::{
-    CollectionStats, MatchExecutionStrategy, MatchQueryPlanner,
-};
-use crate::collection::stats::CollectionStats as CoreCollectionStats;
-use crate::velesql::MatchClause;
-
-/// Query execution plan generated by EXPLAIN.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct QueryPlan {
-    /// Root node of the plan tree.
-    pub root: PlanNode,
-    /// Estimated execution cost in milliseconds.
-    pub estimated_cost_ms: f64,
-    /// Index type used (if any).
-    pub index_used: Option<IndexType>,
-    /// Filter strategy.
-    pub filter_strategy: FilterStrategy,
-    /// WITH clause options surfaced for EXPLAIN output (issue #471).
-    ///
-    /// Each entry is a `(key, value_display)` pair extracted from the parsed
-    /// `WithClause`.  Empty when the query has no WITH clause.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub with_options: Vec<(String, String)>,
-    /// LET bindings surfaced for EXPLAIN output (issue #471).
-    ///
-    /// Each entry is `"name = expr"` in declaration order.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub let_bindings: Vec<String>,
-    /// Fusion strategy info surfaced for EXPLAIN output (issue #471).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub fusion_info: Option<FusionInfo>,
-    /// Whether this plan was served from the compiled plan cache (CACHE-02).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cache_hit: Option<bool>,
-    /// How many times this cached plan has been reused (CACHE-02).
-    ///
-    /// The value is the `reuse_count` atomically incremented inside the cache on every
-    /// successful `get()`. A value of 1 means the plan was found in cache and this is
-    /// its first reuse. The count includes the current call to `explain_query`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub plan_reuse_count: Option<u64>,
-}
-
-/// A node in the query execution plan tree.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum PlanNode {
-    /// Vector similarity search operation.
-    VectorSearch(VectorSearchPlan),
-    /// Metadata filter operation.
-    Filter(FilterPlan),
-    /// Limit results.
-    Limit(LimitPlan),
-    /// Offset skip.
-    Offset(OffsetPlan),
-    /// Table scan (no index).
-    TableScan(TableScanPlan),
-    /// Property index lookup (O(1) instead of scan).
-    IndexLookup(IndexLookupPlan),
-    /// Sequential operations.
-    Sequence(Vec<PlanNode>),
-    /// MATCH graph traversal (EPIC-046 US-004).
-    MatchTraversal(MatchTraversalPlan),
-}
-
-/// Vector search plan details.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct VectorSearchPlan {
-    /// Collection name.
-    pub collection: String,
-    /// `ef_search` parameter (for HNSW).
-    pub ef_search: u32,
-    /// Number of candidates to retrieve.
-    pub candidates: u32,
-}
-
-/// Filter plan details with histogram-based estimation.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct FilterPlan {
-    /// Filter conditions as string representation.
-    pub conditions: String,
-    /// Estimated selectivity (0.0 - 1.0).
-    pub selectivity: f64,
-    /// Estimated number of rows after filtering (selectivity × total_rows).
-    /// Populated when collection stats are available during plan construction.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub estimated_rows: Option<u64>,
-    /// How the estimate was produced: `"histogram"`, `"cardinality"`, or `"no histogram"`.
-    /// Populated during plan construction based on histogram availability.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub estimation_method: Option<String>,
-}
-
-/// Limit plan details.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct LimitPlan {
-    /// Maximum number of results.
-    pub count: u64,
-}
-
-/// Offset plan details.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct OffsetPlan {
-    /// Number of results to skip.
-    pub count: u64,
-}
-
-/// Table scan plan (no index used).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct TableScanPlan {
-    /// Collection name.
-    pub collection: String,
-}
-
-/// Property index lookup plan (O(1) lookup).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct IndexLookupPlan {
-    /// Label being queried.
-    pub label: String,
-    /// Property name with index.
-    pub property: String,
-    /// Value being looked up (as string representation).
-    pub value: String,
-}
-
-/// MATCH traversal plan (EPIC-046 US-004).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct MatchTraversalPlan {
-    /// Execution strategy chosen by planner.
-    pub strategy: String,
-    /// Start node labels.
-    pub start_labels: Vec<String>,
-    /// Maximum traversal depth.
-    pub max_depth: u32,
-    /// Number of relationships in pattern.
-    pub relationship_count: usize,
-    /// Has similarity condition.
-    pub has_similarity: bool,
-    /// Similarity threshold (if any).
-    pub similarity_threshold: Option<f32>,
-}
-
-/// Fusion strategy info for EXPLAIN output (issue #471).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct FusionInfo {
-    /// Strategy name (e.g. "RRF", "Weighted", "Maximum", "RSF").
-    pub strategy: String,
-    /// RRF k parameter (if applicable).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub k: Option<u32>,
-    /// Weights as `"vector=0.7, graph=0.3"` or similar; `None` when unset.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub weights: Option<String>,
-}
-
-/// EXPLAIN output with optional ANALYZE stats and cost model metadata.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExplainOutput {
-    /// The query plan.
-    pub plan: QueryPlan,
-    /// Actual execution statistics (only with ANALYZE).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub actual_stats: Option<ActualStats>,
-    /// Per-node execution statistics (only with ANALYZE).
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub node_stats: Vec<NodeStats>,
-    /// Effective cost factors used for this plan.
-    /// `None` when EXPLAIN is called without ANALYZE.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub cost_factors: Option<OperationCostFactors>,
-    /// Calibration source: `"histogram"` if calibrated, `"default"` otherwise.
-    /// `None` when EXPLAIN is called without ANALYZE.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub calibration_source: Option<String>,
-}
-
-impl ExplainOutput {
-    /// Creates an EXPLAIN-only output (no actual stats, no cost metadata).
-    #[must_use]
-    pub fn plan_only(plan: QueryPlan) -> Self {
-        Self {
-            plan,
-            actual_stats: None,
-            node_stats: vec![],
-            cost_factors: None,
-            calibration_source: None,
-        }
-    }
-
-    /// Creates an EXPLAIN ANALYZE output with actual stats (cost metadata not yet populated).
-    #[must_use]
-    pub fn with_stats(
-        plan: QueryPlan,
-        actual_stats: ActualStats,
-        node_stats: Vec<NodeStats>,
-    ) -> Self {
-        Self {
-            plan,
-            actual_stats: Some(actual_stats),
-            node_stats,
-            cost_factors: None,
-            calibration_source: None,
-        }
-    }
-
-    /// Populates cost factors and calibration source from collection stats.
-    ///
-    /// If `calibrated_cost_factors` is `Some` in the stats, uses those factors
-    /// and sets `calibration_source` to `"histogram"`. Otherwise, uses the
-    /// default factors and sets `calibration_source` to `"default"`.
-    #[must_use]
-    pub fn with_cost_metadata(mut self, stats: &CoreCollectionStats) -> Self {
-        if let Some(ref factors) = stats.calibrated_cost_factors {
-            self.cost_factors = Some(factors.clone());
-            self.calibration_source = Some("histogram".to_string());
-        } else {
-            self.cost_factors = Some(OperationCostFactors::default());
-            self.calibration_source = Some("default".to_string());
-        }
-        self
-    }
-}
-
-/// Actual execution statistics for EXPLAIN ANALYZE.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ActualStats {
-    /// Actual number of rows returned.
-    pub actual_rows: u64,
-    /// Actual execution time in milliseconds.
-    pub actual_time_ms: f64,
-    /// Number of loop iterations.
-    pub loops: u64,
-    /// Number of nodes visited (for graph traversal).
-    pub nodes_visited: u64,
-    /// Number of edges traversed.
-    pub edges_traversed: u64,
-}
-
-/// Per-plan-node **estimated** execution statistics.
-///
-/// All values in this struct are synthetic heuristics derived from the
-/// plan-global `ActualStats::actual_time_ms` using fixed weight fractions
-/// — they are **not** individually measured per node.  The `estimated`
-/// field is always `true` until real per-node instrumentation lands
-/// (#467–#469).
-///
-/// Field names use the `actual_` prefix for forward-compatibility with
-/// the future instrumented variant; the `estimated` flag distinguishes
-/// the two modes at runtime.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct NodeStats {
-    /// Node label (e.g. "VectorSearch", "Filter", "Limit").
-    pub node_label: String,
-    /// Estimated wall-clock time for this node in milliseconds.
-    /// Derived by distributing the total `actual_time_ms` across nodes using
-    /// normalized weight fractions — NOT a real per-node measurement.
-    /// Foundational for the CBO feedback loop; will be replaced by
-    /// instrumented timing (#467–#469).
-    pub actual_time_ms: f64,
-    /// Estimated rows entering this node (heuristic approximation).
-    pub actual_rows_in: u64,
-    /// Estimated rows leaving this node (heuristic approximation).
-    pub actual_rows_out: u64,
-    /// Number of loop iterations (1 for non-looping nodes).
-    pub loops: u64,
-    /// When `true`, `actual_time_ms`, `actual_rows_in` and `actual_rows_out`
-    /// are heuristic estimates derived from fixed proportions, not real
-    /// per-node measurements. Will become `false` once instrumented timing
-    /// lands (#467–#469).
-    pub estimated: bool,
-}
-
-/// Type of index used in the query.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum IndexType {
-    /// HNSW index for vector search.
-    Hnsw,
-    /// Flat index (brute force).
-    Flat,
-    /// Binary quantization index.
-    BinaryQuantization,
-    /// Property index for equality lookups.
-    Property,
-}
-
-/// Strategy for applying filters.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
-#[non_exhaustive]
-pub enum FilterStrategy {
-    /// No filter.
-    #[default]
-    None,
-    /// Pre-filtering: filter before vector search (high selectivity).
-    PreFilter,
-    /// Post-filtering: filter after vector search (low selectivity).
-    PostFilter,
-}
-
-impl QueryPlan {
-    /// Creates a new query plan from a SELECT statement.
-    #[must_use]
-    pub fn from_select(stmt: &SelectStatement) -> Self {
-        Self::from_select_with_indexed_fields(stmt, &HashSet::new())
-    }
-
-    /// Creates a new query plan from SELECT with known indexed metadata fields.
-    #[must_use]
-    pub fn from_select_with_indexed_fields(
-        stmt: &SelectStatement,
-        indexed_fields: &HashSet<String>,
-    ) -> Self {
-        let mut has_vector_search = false;
-        let mut filter_conditions = Vec::new();
-        let mut index_lookup = None;
-
-        if let Some(ref condition) = stmt.where_clause {
-            Self::analyze_condition(condition, &mut has_vector_search, &mut filter_conditions);
-            index_lookup = Self::extract_index_lookup(condition, indexed_fields);
-        }
-
-        let (mut nodes, index_used) = Self::build_scan_node(stmt, has_vector_search, index_lookup);
-        let filter_strategy = Self::append_filter_nodes(&mut nodes, &filter_conditions, stmt);
-
-        let mut plan = Self::assemble_plan(nodes, index_used, filter_strategy, has_vector_search);
-        plan.with_options = Self::extract_with_options(stmt);
-        plan.fusion_info = Self::extract_fusion_info(stmt);
-        plan
-    }
-
-    /// Creates a full query plan from a `Query`, including LET bindings (issue #471).
-    #[must_use]
-    pub fn from_query(query: &super::ast::Query) -> Self {
-        let mut plan = Self::from_select(&query.select);
-        plan.let_bindings = Self::format_let_bindings(&query.let_bindings);
-        plan
-    }
-
-    /// Collapses a `Vec<PlanNode>` into a single root, estimates cost, and builds the plan.
-    fn assemble_plan(
-        mut nodes: Vec<PlanNode>,
-        index_used: Option<IndexType>,
-        filter_strategy: FilterStrategy,
-        has_vector_search: bool,
-    ) -> Self {
-        let root = if nodes.len() == 1 {
-            nodes.swap_remove(0)
-        } else {
-            PlanNode::Sequence(nodes)
-        };
-        let estimated_cost_ms = Self::estimate_cost(&root, has_vector_search);
-        Self {
-            root,
-            estimated_cost_ms,
-            index_used,
-            filter_strategy,
-            with_options: Vec::new(),
-            let_bindings: Vec::new(),
-            fusion_info: None,
-            cache_hit: None,
-            plan_reuse_count: None,
-        }
-    }
-
-    /// Default `ef_search` when the WITH clause does not specify one.
-    const DEFAULT_EF_SEARCH: u32 = 100;
-
-    /// Builds the primary scan node based on search type.
-    fn build_scan_node(
-        stmt: &SelectStatement,
-        has_vector_search: bool,
-        index_lookup: Option<(String, String)>,
-    ) -> (Vec<PlanNode>, Option<IndexType>) {
-        let mut nodes = Vec::new();
-        let index_used;
-
-        if has_vector_search {
-            index_used = Some(IndexType::Hnsw);
-            let candidates = u32::try_from(stmt.limit.unwrap_or(50)).unwrap_or(u32::MAX);
-            let ef_search = Self::resolve_ef_search(stmt);
-            nodes.push(PlanNode::VectorSearch(VectorSearchPlan {
-                collection: stmt.from.clone(),
-                ef_search,
-                candidates,
-            }));
-        } else if let Some((property, value)) = index_lookup {
-            index_used = Some(IndexType::Property);
-            nodes.push(PlanNode::IndexLookup(IndexLookupPlan {
-                label: stmt.from.clone(),
-                property,
-                value,
-            }));
-        } else {
-            index_used = None;
-            nodes.push(PlanNode::TableScan(TableScanPlan {
-                collection: stmt.from.clone(),
-            }));
-        }
-
-        (nodes, index_used)
-    }
-
-    /// Reads `ef_search` from the WITH clause, falling back to [`Self::DEFAULT_EF_SEARCH`].
-    #[allow(clippy::cast_possible_truncation)]
-    fn resolve_ef_search(stmt: &SelectStatement) -> u32 {
-        stmt.with_clause
-            .as_ref()
-            .and_then(super::ast::WithClause::get_ef_search)
-            .map_or(Self::DEFAULT_EF_SEARCH, |v| v as u32)
-    }
-
-    /// Extracts WITH clause options as display pairs (issue #471).
-    fn extract_with_options(stmt: &SelectStatement) -> Vec<(String, String)> {
-        let Some(ref wc) = stmt.with_clause else {
-            return Vec::new();
-        };
-        wc.options
-            .iter()
-            .map(|opt| (opt.key.clone(), formatter::format_with_value(&opt.value)))
-            .collect()
-    }
-
-    /// Extracts FUSION clause info for EXPLAIN display (issue #471).
-    fn extract_fusion_info(stmt: &SelectStatement) -> Option<FusionInfo> {
-        let fc = stmt.fusion_clause.as_ref()?;
-        let strategy = match fc.strategy {
-            super::ast::FusionStrategyType::Rrf => "RRF",
-            super::ast::FusionStrategyType::Weighted => "Weighted",
-            super::ast::FusionStrategyType::Maximum => "Maximum",
-            super::ast::FusionStrategyType::Rsf => "RSF",
-            super::ast::FusionStrategyType::Average => "Average",
-        };
-        let weights = Self::format_fusion_weights(fc);
-        Some(FusionInfo {
-            strategy: strategy.to_string(),
-            k: fc.k,
-            weights,
-        })
-    }
-
-    /// Formats fusion weights into a human-readable string.
-    fn format_fusion_weights(fc: &super::ast::FusionClause) -> Option<String> {
-        let mut parts = Vec::new();
-        if let Some(vw) = fc.vector_weight {
-            parts.push(format!("vector={vw}"));
-        }
-        if let Some(gw) = fc.graph_weight {
-            parts.push(format!("graph={gw}"));
-        }
-        if let Some(dw) = fc.dense_weight {
-            parts.push(format!("dense={dw}"));
-        }
-        if let Some(sw) = fc.sparse_weight {
-            parts.push(format!("sparse={sw}"));
-        }
-        if parts.is_empty() {
-            None
-        } else {
-            Some(parts.join(", "))
-        }
-    }
-
-    /// Formats LET bindings as `"name = expr"` strings (issue #471).
-    fn format_let_bindings(bindings: &[LetBinding]) -> Vec<String> {
-        bindings
-            .iter()
-            .map(|b| format!("{} = {}", b.name, b.expr))
-            .collect()
-    }
-
-    /// Appends filter, offset, and limit nodes; returns the filter strategy.
-    fn append_filter_nodes(
-        nodes: &mut Vec<PlanNode>,
-        filter_conditions: &[String],
-        stmt: &SelectStatement,
-    ) -> FilterStrategy {
-        let mut filter_strategy = FilterStrategy::None;
-
-        if !filter_conditions.is_empty() {
-            let selectivity = Self::estimate_selectivity(filter_conditions);
-            filter_strategy = if selectivity > 0.1 {
-                FilterStrategy::PostFilter
-            } else {
-                FilterStrategy::PreFilter
-            };
-            nodes.push(PlanNode::Filter(FilterPlan {
-                conditions: filter_conditions.join(" AND "),
-                selectivity,
-                estimated_rows: None,
-                estimation_method: None,
-            }));
-        }
-
-        if let Some(offset) = stmt.offset {
-            nodes.push(PlanNode::Offset(OffsetPlan { count: offset }));
-        }
-        if let Some(limit) = stmt.limit {
-            nodes.push(PlanNode::Limit(LimitPlan { count: limit }));
-        }
-
-        filter_strategy
-    }
-
-    /// Analyzes a condition to extract vector search and filter info.
-    fn analyze_condition(
-        condition: &Condition,
-        has_vector_search: &mut bool,
-        filter_conditions: &mut Vec<String>,
-    ) {
-        match condition {
-            Condition::VectorSearch(_)
-            | Condition::VectorFusedSearch(_)
-            | Condition::SparseVectorSearch(_)
-            | Condition::Similarity(_) => {
-                *has_vector_search = true;
-            }
-            Condition::Comparison(cmp) => {
-                filter_conditions.push(format!("{} {} ?", cmp.column, cmp.operator.as_str()));
-            }
-            Condition::In(inc) => {
-                let op = if inc.negated { "NOT IN" } else { "IN" };
-                filter_conditions.push(format!("{} {op} (...)", inc.column));
-            }
-            Condition::Between(btw) => {
-                filter_conditions.push(format!("{} BETWEEN ? AND ?", btw.column));
-            }
-            Condition::Like(lk) => {
-                filter_conditions.push(format!("{} LIKE ?", lk.column));
-            }
-            Condition::IsNull(isn) => {
-                let op = if isn.is_null {
-                    "IS NULL"
-                } else {
-                    "IS NOT NULL"
-                };
-                filter_conditions.push(format!("{} {op}", isn.column));
-            }
-            Condition::Match(m) => {
-                filter_conditions.push(format!("{} MATCH ?", m.column));
-            }
-            Condition::ContainsText(ct) => {
-                filter_conditions.push(format!("{} CONTAINS_TEXT ?", ct.column));
-            }
-            Condition::GraphMatch(_) => {
-                filter_conditions.push("MATCH (...)".to_string());
-            }
-            Condition::Contains(cc) => {
-                let mode_str = match cc.mode {
-                    crate::velesql::ContainsMode::Single => "CONTAINS",
-                    crate::velesql::ContainsMode::Any => "CONTAINS ANY",
-                    crate::velesql::ContainsMode::All => "CONTAINS ALL",
-                };
-                filter_conditions.push(format!("{} {mode_str} ?", cc.column));
-            }
-            Condition::GeoDistance(gd) => {
-                filter_conditions.push(format!(
-                    "GEO_DISTANCE({}, {}, {}) {} ?",
-                    gd.column,
-                    gd.lat,
-                    gd.lng,
-                    gd.operator.as_str()
-                ));
-            }
-            Condition::GeoBbox(gb) => {
-                filter_conditions.push(format!("GEO_BBOX({}, ...)", gb.column));
-            }
-            Condition::And(left, right) | Condition::Or(left, right) => {
-                Self::analyze_condition(left, has_vector_search, filter_conditions);
-                Self::analyze_condition(right, has_vector_search, filter_conditions);
-            }
-            Condition::Not(inner) | Condition::Group(inner) => {
-                Self::analyze_condition(inner, has_vector_search, filter_conditions);
-            }
-        }
-    }
-
-    fn extract_index_lookup(
-        condition: &Condition,
-        indexed_fields: &HashSet<String>,
-    ) -> Option<(String, String)> {
-        if let Condition::Comparison(cmp) = condition {
-            if cmp.operator == crate::velesql::CompareOp::Eq && indexed_fields.contains(&cmp.column)
-            {
-                return Some((cmp.column.clone(), format!("{:?}", cmp.value)));
-            }
-        }
-        if let Condition::In(inc) = condition {
-            if indexed_fields.contains(&inc.column) {
-                let op = if inc.negated { "NOT IN" } else { "IN" };
-                return Some((inc.column.clone(), format!("{op} (...)")));
-            }
-        }
-        None
-    }
-
-    /// Estimates selectivity (placeholder - would need statistics in production).
-    pub(crate) fn estimate_selectivity(conditions: &[String]) -> f64 {
-        node_stats::estimate_selectivity(conditions)
-    }
-
-    /// Estimates execution cost in milliseconds.
-    fn estimate_cost(root: &PlanNode, has_vector_search: bool) -> f64 {
-        node_stats::estimate_cost(root, has_vector_search)
-    }
-
-    /// Returns the heuristic cost for a single plan node.
-    #[cfg(test)]
-    pub(crate) fn node_cost(node: &PlanNode) -> f64 {
-        node_stats::node_cost(node)
-    }
-
-    /// Creates a new query plan from a MATCH clause (EPIC-046 US-004).
-    #[must_use]
-    pub fn from_match(match_clause: &MatchClause, stats: &CollectionStats) -> Self {
-        let strategy = MatchQueryPlanner::plan(match_clause, stats);
-        let strategy_explanation = MatchQueryPlanner::explain(&strategy);
-
-        let (start_labels, max_depth, has_similarity, similarity_threshold) =
-            Self::extract_strategy_info(&strategy);
-
-        let relationship_count = match_clause
-            .patterns
-            .first()
-            .map_or(0, |p| p.relationships.len());
-
-        let traversal = PlanNode::MatchTraversal(MatchTraversalPlan {
-            strategy: strategy_explanation,
-            start_labels,
-            max_depth,
-            relationship_count,
-            has_similarity,
-            similarity_threshold,
-        });
-
-        let mut nodes = vec![traversal];
-        if let Some(limit) = match_clause.return_clause.limit {
-            nodes.push(PlanNode::Limit(LimitPlan { count: limit }));
-        }
-
-        let index_used = if has_similarity {
-            Some(IndexType::Hnsw)
-        } else {
-            None
-        };
-
-        Self::assemble_plan(nodes, index_used, FilterStrategy::None, has_similarity)
-    }
-
-    /// Extracts traversal parameters from a `MatchExecutionStrategy`.
-    fn extract_strategy_info(
-        strategy: &MatchExecutionStrategy,
-    ) -> (Vec<String>, u32, bool, Option<f32>) {
-        match strategy {
-            MatchExecutionStrategy::GraphFirst {
-                start_labels,
-                max_depth,
-            } => (start_labels.clone(), *max_depth, false, None),
-            MatchExecutionStrategy::VectorFirst { threshold, .. } => {
-                (Vec::new(), 1, true, Some(*threshold))
-            }
-            MatchExecutionStrategy::Parallel {
-                graph_hint,
-                vector_hint,
-            } => {
-                let (labels, depth) = match graph_hint.as_ref() {
-                    MatchExecutionStrategy::GraphFirst {
-                        start_labels,
-                        max_depth,
-                    } => (start_labels.clone(), *max_depth),
-                    _ => (Vec::new(), 1),
-                };
-                let threshold = match vector_hint.as_ref() {
-                    MatchExecutionStrategy::VectorFirst { threshold, .. } => Some(*threshold),
-                    _ => None,
-                };
-                (labels, depth, true, threshold)
-            }
-        }
-    }
-}
-
-// Rendering, formatting, Display impl, and as_str() methods moved to explain/formatter.rs
-// Node-stats helpers (build_leaf_node_stats, etc.) moved to explain/node_stats.rs
-// Tests in explain_tests.rs per project rules (tests in separate files)
