+++ crates/velesdb-core/src/velesql/explain/types.rs
@@ -1,0 +1,302 @@
+//! Type definitions for `VelesQL` EXPLAIN query plans.
+//!
+//! Contains all structs and enums used by the explain module:
+//! `QueryPlan`, `PlanNode`, `ExplainOutput`, `ActualStats`, `NodeStats`, etc.
+
+use serde::{Deserialize, Serialize};
+
+use crate::collection::query_cost::cost_model::OperationCostFactors;
+use crate::collection::stats::CollectionStats as CoreCollectionStats;
+
+/// Query execution plan generated by EXPLAIN.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct QueryPlan {
+    /// Root node of the plan tree.
+    pub root: PlanNode,
+    /// Estimated execution cost in milliseconds.
+    pub estimated_cost_ms: f64,
+    /// Index type used (if any).
+    pub index_used: Option<IndexType>,
+    /// Filter strategy.
+    pub filter_strategy: FilterStrategy,
+    /// WITH clause options surfaced for EXPLAIN output (issue #471).
+    ///
+    /// Each entry is a `(key, value_display)` pair extracted from the parsed
+    /// `WithClause`.  Empty when the query has no WITH clause.
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub with_options: Vec<(String, String)>,
+    /// LET bindings surfaced for EXPLAIN output (issue #471).
+    ///
+    /// Each entry is `"name = expr"` in declaration order.
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub let_bindings: Vec<String>,
+    /// Fusion strategy info surfaced for EXPLAIN output (issue #471).
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub fusion_info: Option<FusionInfo>,
+    /// Whether this plan was served from the compiled plan cache (CACHE-02).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub cache_hit: Option<bool>,
+    /// How many times this cached plan has been reused (CACHE-02).
+    ///
+    /// The value is the `reuse_count` atomically incremented inside the cache on every
+    /// successful `get()`. A value of 1 means the plan was found in cache and this is
+    /// its first reuse. The count includes the current call to `explain_query`.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub plan_reuse_count: Option<u64>,
+}
+
+/// A node in the query execution plan tree.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[non_exhaustive]
+pub enum PlanNode {
+    /// Vector similarity search operation.
+    VectorSearch(VectorSearchPlan),
+    /// Metadata filter operation.
+    Filter(FilterPlan),
+    /// Limit results.
+    Limit(LimitPlan),
+    /// Offset skip.
+    Offset(OffsetPlan),
+    /// Table scan (no index).
+    TableScan(TableScanPlan),
+    /// Property index lookup (O(1) instead of scan).
+    IndexLookup(IndexLookupPlan),
+    /// Sequential operations.
+    Sequence(Vec<PlanNode>),
+    /// MATCH graph traversal (EPIC-046 US-004).
+    MatchTraversal(MatchTraversalPlan),
+}
+
+/// Vector search plan details.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct VectorSearchPlan {
+    /// Collection name.
+    pub collection: String,
+    /// `ef_search` parameter (for HNSW).
+    pub ef_search: u32,
+    /// Number of candidates to retrieve.
+    pub candidates: u32,
+}
+
+/// Filter plan details with histogram-based estimation.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct FilterPlan {
+    /// Filter conditions as string representation.
+    pub conditions: String,
+    /// Estimated selectivity (0.0 - 1.0).
+    pub selectivity: f64,
+    /// Estimated number of rows after filtering (selectivity * total_rows).
+    /// Populated when collection stats are available during plan construction.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub estimated_rows: Option<u64>,
+    /// How the estimate was produced: `"histogram"`, `"cardinality"`, or `"no histogram"`.
+    /// Populated during plan construction based on histogram availability.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub estimation_method: Option<String>,
+}
+
+/// Limit plan details.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct LimitPlan {
+    /// Maximum number of results.
+    pub count: u64,
+}
+
+/// Offset plan details.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct OffsetPlan {
+    /// Number of results to skip.
+    pub count: u64,
+}
+
+/// Table scan plan (no index used).
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct TableScanPlan {
+    /// Collection name.
+    pub collection: String,
+}
+
+/// Property index lookup plan (O(1) lookup).
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct IndexLookupPlan {
+    /// Label being queried.
+    pub label: String,
+    /// Property name with index.
+    pub property: String,
+    /// Value being looked up (as string representation).
+    pub value: String,
+}
+
+/// MATCH traversal plan (EPIC-046 US-004).
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct MatchTraversalPlan {
+    /// Execution strategy chosen by planner.
+    pub strategy: String,
+    /// Start node labels.
+    pub start_labels: Vec<String>,
+    /// Maximum traversal depth.
+    pub max_depth: u32,
+    /// Number of relationships in pattern.
+    pub relationship_count: usize,
+    /// Has similarity condition.
+    pub has_similarity: bool,
+    /// Similarity threshold (if any).
+    pub similarity_threshold: Option<f32>,
+}
+
+/// Fusion strategy info for EXPLAIN output (issue #471).
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct FusionInfo {
+    /// Strategy name (e.g. "RRF", "Weighted", "Maximum", "RSF").
+    pub strategy: String,
+    /// RRF k parameter (if applicable).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub k: Option<u32>,
+    /// Weights as `"vector=0.7, graph=0.3"` or similar; `None` when unset.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub weights: Option<String>,
+}
+
+/// EXPLAIN output with optional ANALYZE stats and cost model metadata.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ExplainOutput {
+    /// The query plan.
+    pub plan: QueryPlan,
+    /// Actual execution statistics (only with ANALYZE).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub actual_stats: Option<ActualStats>,
+    /// Per-node execution statistics (only with ANALYZE).
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub node_stats: Vec<NodeStats>,
+    /// Effective cost factors used for this plan.
+    /// `None` when EXPLAIN is called without ANALYZE.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub cost_factors: Option<OperationCostFactors>,
+    /// Calibration source: `"histogram"` if calibrated, `"default"` otherwise.
+    /// `None` when EXPLAIN is called without ANALYZE.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub calibration_source: Option<String>,
+}
+
+impl ExplainOutput {
+    /// Creates an EXPLAIN-only output (no actual stats, no cost metadata).
+    #[must_use]
+    pub fn plan_only(plan: QueryPlan) -> Self {
+        Self {
+            plan,
+            actual_stats: None,
+            node_stats: vec![],
+            cost_factors: None,
+            calibration_source: None,
+        }
+    }
+
+    /// Creates an EXPLAIN ANALYZE output with actual stats (cost metadata not yet populated).
+    #[must_use]
+    pub fn with_stats(
+        plan: QueryPlan,
+        actual_stats: ActualStats,
+        node_stats: Vec<NodeStats>,
+    ) -> Self {
+        Self {
+            plan,
+            actual_stats: Some(actual_stats),
+            node_stats,
+            cost_factors: None,
+            calibration_source: None,
+        }
+    }
+
+    /// Populates cost factors and calibration source from collection stats.
+    ///
+    /// If `calibrated_cost_factors` is `Some` in the stats, uses those factors
+    /// and sets `calibration_source` to `"histogram"`. Otherwise, uses the
+    /// default factors and sets `calibration_source` to `"default"`.
+    #[must_use]
+    pub fn with_cost_metadata(mut self, stats: &CoreCollectionStats) -> Self {
+        if let Some(ref factors) = stats.calibrated_cost_factors {
+            self.cost_factors = Some(factors.clone());
+            self.calibration_source = Some("histogram".to_string());
+        } else {
+            self.cost_factors = Some(OperationCostFactors::default());
+            self.calibration_source = Some("default".to_string());
+        }
+        self
+    }
+}
+
+/// Actual execution statistics for EXPLAIN ANALYZE.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ActualStats {
+    /// Actual number of rows returned.
+    pub actual_rows: u64,
+    /// Actual execution time in milliseconds.
+    pub actual_time_ms: f64,
+    /// Number of loop iterations.
+    pub loops: u64,
+    /// Number of nodes visited (for graph traversal).
+    pub nodes_visited: u64,
+    /// Number of edges traversed.
+    pub edges_traversed: u64,
+}
+
+/// Per-plan-node **estimated** execution statistics.
+///
+/// All values in this struct are synthetic heuristics derived from the
+/// plan-global `ActualStats::actual_time_ms` using fixed weight fractions
+/// -- they are **not** individually measured per node.  The `estimated`
+/// field is always `true` until real per-node instrumentation lands
+/// (#467-#469).
+///
+/// Field names use the `actual_` prefix for forward-compatibility with
+/// the future instrumented variant; the `estimated` flag distinguishes
+/// the two modes at runtime.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct NodeStats {
+    /// Node label (e.g. "VectorSearch", "Filter", "Limit").
+    pub node_label: String,
+    /// Estimated wall-clock time for this node in milliseconds.
+    /// Derived by distributing the total `actual_time_ms` across nodes using
+    /// normalized weight fractions -- NOT a real per-node measurement.
+    /// Foundational for the CBO feedback loop; will be replaced by
+    /// instrumented timing (#467-#469).
+    pub actual_time_ms: f64,
+    /// Estimated rows entering this node (heuristic approximation).
+    pub actual_rows_in: u64,
+    /// Estimated rows leaving this node (heuristic approximation).
+    pub actual_rows_out: u64,
+    /// Number of loop iterations (1 for non-looping nodes).
+    pub loops: u64,
+    /// When `true`, `actual_time_ms`, `actual_rows_in` and `actual_rows_out`
+    /// are heuristic estimates derived from fixed proportions, not real
+    /// per-node measurements. Will become `false` once instrumented timing
+    /// lands (#467-#469).
+    pub estimated: bool,
+}
+
+/// Type of index used in the query.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[non_exhaustive]
+pub enum IndexType {
+    /// HNSW index for vector search.
+    Hnsw,
+    /// Flat index (brute force).
+    Flat,
+    /// Binary quantization index.
+    BinaryQuantization,
+    /// Property index for equality lookups.
+    Property,
+}
+
+/// Strategy for applying filters.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
+#[non_exhaustive]
+pub enum FilterStrategy {
+    /// No filter.
+    #[default]
+    None,
+    /// Pre-filtering: filter before vector search (high selectivity).
+    PreFilter,
+    /// Post-filtering: filter after vector search (low selectivity).
+    PostFilter,
+}
